+++ ios/iosremote/SWRevealViewController/SWRevealViewController.m
@@ -90,14 +90,14 @@ typedef enum
     CGFloat moveX = nowPoint.x - _init.x;
     CGFloat moveY = nowPoint.y - _init.y;
     
-    if (fabsf(moveX) > kDirectionPanThreshold)
+    if (fabs(moveX) > kDirectionPanThreshold)
     {
         if (_direction == SWDirectionPanGestureRecognizerHorizontal)
             _dragging = YES;
         else
             self.state = UIGestureRecognizerStateFailed;
     }
-    else if (fabsf(moveY) > kDirectionPanThreshold)
+    else if (fabs(moveY) > kDirectionPanThreshold)
     {
         if (_direction == SWDirectionPanGestureRecognizerVertical)
             _dragging = YES ;
@@ -871,7 +871,7 @@ static NSString * const SWSegueRightIdentifier = @"sw_right";
     NSTimeInterval duration = _toggleAnimationDuration;
 
     // Velocity driven change:
-    if (fabsf(velocity) > _quickFlickVelocity)
+    if (fabs(velocity) > _quickFlickVelocity)
     {
         // we may need to set the drag position and to adjust the animation duration
         CGFloat journey = xPosition;
@@ -889,7 +889,7 @@ static NSString * const SWSegueRightIdentifier = @"sw_right";
             }
         }
         
-        duration = fabsf(journey/velocity);
+        duration = fabs(journey/velocity);
     }
     
     // Position driven change:
